update some dependencies
