+++ frontend/app/(signInsignup)/(ngo)/thirdpage/index.tsx
@@ -1,12 +1,11 @@
-import ButtonGoogle from "@/components/button-google"; // Unused import
 import ButtonSuccess from "@/components/button-success";
-import ButtonText from "@/components/button-text"; // Unused import
+import ButtonText from "@/components/button-text"; 
 import TextInputStyled from "@/components/text-input";
 import { router } from "expo-router";
 import { useRef, useState } from "react";
 import { StyleSheet, View } from "react-native";
 import { CheckBox } from "react-native-elements/dist/checkbox/CheckBox";
-import { TextInput } from "react-native-gesture-handler"; // Ensure correct import
+import { TextInput } from "react-native-gesture-handler"; 
 
 const Third = () => {
   const [email, setEmail] = useState("");
