+++ frontend/package.json
@@ -38,10 +38,12 @@
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-native": "0.76.2",
+    "react-native-elements": "^3.4.3",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.1.0",
+    "react-native-vector-icons": "^10.2.0",
     "react-native-web": "~0.19.13",
     "react-native-webview": "13.12.2"
   },
